requirements.txt
streamlit==1.32.2
google-genai
python-dotenv